Added working version of extended DrMon with SiPMs
Copied old files where I started to work on a monitoring with SiPMs included for DrMon
